backend/requirements.txt
# 🚀 VibeTalk Backend - Production Dependencies
# Developer: Yash Ankush Mishra (Rangra Developer)

# Django Core
Django==5.0.1
djangorestframework==3.14.0

# Database
dj-database-url==2.1.0
psycopg2-binary==2.9.9

# CORS & Security
django-cors-headers==4.3.1
python-dotenv==1.0.0

# WebSocket Support
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.0

# Caching
redis==5.0.1
django-redis==5.4.0

# Authentication
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Media Storage (Optional - for S3)
boto3==1.34.27
django-storages==1.14.2

# Monitoring & Logging
sentry-sdk==1.39.2

# Rate Limiting
django-ratelimit==4.1.0

# Image Processing
Pillow==10.2.0

# Utils
pytz==2023.3
requests==2.31.0

# Development Tools (optional)
django-debug-toolbar==4.2.0
